Eco Visual/backend/requirements.txt
ultralytics
fastapi
uvicorn
python-multipart
opencv-python
numpy
pillow
